requirements.txt
pymongo==4.6.1
python-dotenv>=0.19.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
